common: split Wayland socket setup to separate file
More complicated setup logic will be added next commit.
