+++ common/Makefile.am.inc
@@ -173,6 +173,8 @@ libflatpak_common_la_SOURCES = \
 	common/flatpak-run-pulseaudio.c \
 	common/flatpak-run-sockets-private.h \
 	common/flatpak-run-sockets.c \
+	common/flatpak-run-wayland-private.h \
+	common/flatpak-run-wayland.c \
 	common/flatpak-run-x11-private.h \
 	common/flatpak-run-x11.c \
 	common/flatpak-syscalls-private.h \
